chore: move src/ -> internal/
Move code from src/ to internal/, since the code is only meant for use
within this module, per https://go.dev/doc/go1.4#internalpackages.
