+++ generator/internal/genclient/language/internal/rust/rust.go
@@ -12,13 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package golang
+package rust
 
 import (
 	"log/slog"
-	"strings"
 
-	"github.com/googleapis/google-cloud-rust/generator/src/genclient"
+	"github.com/googleapis/google-cloud-rust/generator/internal/genclient"
 	"github.com/iancoleman/strcase"
 )
 
@@ -30,14 +29,12 @@ type Codec struct{}
 
 func (c *Codec) LoadWellKnownTypes(s *genclient.APIState) {
 	timestamp := &genclient.Message{
-		ID:      ".google.protobuf.Timestamp",
-		Name:    "Time",
-		Package: "time",
+		ID:   ".google.protobuf.Timestamp",
+		Name: "String",
 	}
 	duration := &genclient.Message{
-		ID:      ".google.protobuf.Duration",
-		Name:    "Duration",
-		Package: "time",
+		ID:   ".google.protobuf.Duration",
+		Name: "String",
 	}
 	s.MessageByID[timestamp.ID] = timestamp
 	s.MessageByID[duration.ID] = duration
@@ -47,15 +44,15 @@ func (c *Codec) FieldType(f *genclient.Field, state *genclient.APIState) string 
 	var out string
 	switch f.Typez {
 	case genclient.STRING_TYPE:
-		out = "string"
+		out = "String"
 	case genclient.INT64_TYPE:
-		out = "int64"
+		out = "i64"
 	case genclient.INT32_TYPE:
-		out = "int32"
+		out = "i32"
 	case genclient.BOOL_TYPE:
 		out = "bool"
 	case genclient.BYTES_TYPE:
-		out = "[]byte"
+		out = "bytes::Bytes"
 	case genclient.MESSAGE_TYPE:
 		m, ok := state.MessageByID[f.TypezID]
 		if !ok {
@@ -65,10 +62,10 @@ func (c *Codec) FieldType(f *genclient.Field, state *genclient.APIState) string 
 		if m.IsMap {
 			key := c.FieldType(m.Fields[0], state)
 			val := c.FieldType(m.Fields[1], state)
-			out = "map[" + key + "]" + val
+			out = "Option<std::collections::HashMap<" + key + "," + val + ">>"
 			break
 		}
-		out = "*" + c.MessageName(m, state)
+		out = "Option<" + c.MessageName(m, state) + ">"
 	case genclient.ENUM_TYPE:
 		e, ok := state.EnumByID[f.TypezID]
 		if !ok {
@@ -83,7 +80,7 @@ func (c *Codec) FieldType(f *genclient.Field, state *genclient.APIState) string 
 }
 
 func (c *Codec) TemplateDir() string {
-	return "go"
+	return "rust"
 }
 
 func (c *Codec) MethodInOutTypeName(id string, s *genclient.APIState) string {
@@ -117,9 +114,9 @@ func (c *Codec) EnumName(e *genclient.Enum, state *genclient.APIState) string {
 
 func (c *Codec) EnumValueName(e *genclient.EnumValue, state *genclient.APIState) string {
 	if e.Parent.Parent != nil {
-		return c.MessageName(e.Parent.Parent, state) + "_" + strings.ToUpper(e.Name)
+		return c.MessageName(e.Parent.Parent, state) + "_" + strcase.ToCamel(e.Name)
 	}
-	return strings.ToUpper(e.Name)
+	return strcase.ToCamel(e.Name)
 }
 
 func (c *Codec) BodyAccessor(m *genclient.Method, state *genclient.APIState) string {
@@ -127,19 +124,17 @@ func (c *Codec) BodyAccessor(m *genclient.Method, state *genclient.APIState) str
 		// no accessor needed, use the whole request
 		return ""
 	}
-	return "." + strcase.ToCamel(m.HTTPInfo.Body)
+	return "." + strcase.ToSnake(m.HTTPInfo.Body)
 }
 
 func (c *Codec) HTTPPathFmt(m *genclient.HTTPInfo, state *genclient.APIState) string {
-	return genclient.HTTPPathVarRegex.ReplaceAllStringFunc(m.RawPath, func(s string) string { return "%s" })
+	return genclient.HTTPPathVarRegex.ReplaceAllStringFunc(m.RawPath, func(s string) string { return "{}" })
 }
-
 func (c *Codec) HTTPPathArgs(h *genclient.HTTPInfo, state *genclient.APIState) []string {
 	var args []string
 	rawArgs := h.PathArgs()
 	for _, arg := range rawArgs {
-		// TODO(codyoss): https://github.com/googleapis/google-cloud-rust/issues/34
-		args = append(args, ", req."+strcase.ToCamel(arg))
+		args = append(args, "req."+strcase.ToSnake(arg))
 	}
 	return args
 }
@@ -155,7 +150,7 @@ func (c *Codec) QueryParams(m *genclient.Method, state *genclient.APIState) []*g
 	var queryParams []*genclient.Pair
 	for _, field := range msg.Fields {
 		if field.JSONName != "" && !notQuery[field.JSONName] {
-			queryParams = append(queryParams, &genclient.Pair{Key: field.JSONName, Value: "req." + strcase.ToCamel(field.JSONName)})
+			queryParams = append(queryParams, &genclient.Pair{Key: field.JSONName, Value: "req." + strcase.ToSnake(field.JSONName) + ".as_str()"})
 		}
 	}
 	return queryParams
