+++ generator/internal/genclient/model.go
