+++ generator/internal/genclient/templatedata.go
